requirements.txt
fastapi
uvicorn
flet
flet-web
pydantic
pydantic-settings
supabase
google-generativeai
git+https://github.com/miaucl/cookidoo-api.git
python-dotenv
httpx
